refactor participation for CC
